requirements.txt
discord.py
youtube-dl
discordSuperUtils